Genre Dropdown
Added a genre dropdown to the header and put each genre into it. Mostly did HTML and CSS for now, will have to get around to creating new HTML pages when each genre is clicked. The functionality I am thinking of is when a genre is clicked from the dropdown a new page is loaded and sorts that genre by descending popularity.