Many changes ahead
Created the async function getAPIs() which will now take in the discover api and the genre api. This will be used in order to compare the genre ids from each movie with the id's of each corresponding genre in order to categorize them properly. Kind of gave up on the idea of making multiple pages for each genre.